# chat-server/src/types/requests/user.request.ts
export interface saveReqUser {
  email:string;
}
export interface updateReqUser {
  _id: string;
  email:string;
}

export interface deleteReqUser {
  _id: string;
}
export interface getReqUser {
  _id: string;
}